Move "Camera" and "Gallery" actions out of the "Select image" submenu
Probably, most users come to the "Add" screen of Image Occlusion note type screen
just in order to use the "Camera" or the "Gallery" feature;
moving these actions from the submenu to the top screen of the note editor is practical.

Not only from the pragmatic point of view, but also from a theoretical standpoint,
"Gallery" is equivalent to "Select image" in the desktop version,
so "Camera" should be a separate option on the same level with
"Gallery" and "Paste image from clipboard".
In fact, in AnkiMobile, the three options are placed in the same hierarchy.

(Besides, there is more than enough space on the current "Add" screen
for Image Occlusion note type (the screen doesn't show deck selector,
tag field, toolbar, and keyboard); adding more parts will rather improve
the visual balance.)